// frontend/src/api/generated/model/userStatus.ts
export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  blocked: 'blocked',
  archived: 'archived',
} as const;